+++ src/main/java/org/springblade/modules/EnterpriseResource/vo/AptitudeVO.java
@@ -52,7 +52,7 @@ public class AptitudeVO extends Aptitude {
 	 * 证书类别ID
 	 */
 	@ApiModelProperty(value = "证书类别ID")
-	private Integer certificateType;
+	private Long certificateType;
 	@ApiModelProperty(value = "证书类别名称")
 	private String certificateTypeName;
 
@@ -60,7 +60,7 @@ public class AptitudeVO extends Aptitude {
 	 * 等级属性ID
 	 */
 	@ApiModelProperty(value = "等级属性ID")
-	private Integer classType;
+	private Long classType;
 	@ApiModelProperty(value = "等级属性名称")
 	private String classTypeName;
